remove api 파일 명 변경
