+++ src/pages/CommunityDetailPage/components/CommentForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { commentCreateApi } from '../../../apis/commentCreateApi';
+import { commentCreateApi } from '../../../apis/commentApi';
 import { getCookie } from '../../../utils/cookie'; 
 
 const CommentForm = ({ postId, onCommentCreated }) => {
